Move all jsonrpc to shared base provider
